Move plot_countries() in plot.py
